Allow reading request body before sending response headers
When handing a request with data (eg: POST) it's often necessary to read
the body of the request before determining the appropriate HTTP status
code for the response.

Make `ConnectRequest.recv_stream` public so it can be read before
calling `ConnectRequest.respond_with()`.

(`send_stream` is left private so `respond_with()` can ensure it sends
the response's headers before the response body.)